requirements.txt
flatdict
gdown==5.2.0
gitdb==4.0.11
hydra-core
matplotlib
pytorch-lightning
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
timm==1.0.9
tokenizers==0.19.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==1.1.1
torchvision==0.15.2+cu118
tqdm
transformers
wandb
pillow
